ghc: add debian binary bootstrap method
Debian provides .deb binaries for many more platforms than are
available via upstream haskell.org. So we prefer the upstream binaries
when supported on the build system, and otherwise attempt to use Debian
binaries before failing.

This is currently only set up for powerpc64le, but is hopefully set up
in a sufficiently generic way that it can be extended in the future to
other platforms supported by Debian (such as s390x and MIPS).